Limit commit list to the release branch
This stops the tool from looking past the branching point of the
previous release branch. The branching point between current i.e
`release-4.19` and previous `release-4.18` branches is found and used as
a limit in case the previous release tag cannot be found.

If the repository does not have the release-4.y naming structure for
branches, the existing behavior is maintained.